Update rpn_generator.py
Added lines like
"    print("im_proposals - boxes: ", boxes)
    print("im_proposals - scores: ", scores)"
to every function.

Hopefully now, when I run the demo code on an image, it will print out the boxes and scores for each picture. Not really sure how this will work, but easy to remove if it brings nothing exciting or brings any failures.